Move torch version to another file, and get image input as float32 dtype
