+++ chromakey/__init__.py
@@ -1,3 +1,3 @@
-from .core import chroma_key, chroma_key_vectorized
+from .core import chroma_key
 
-__all__ = ["chroma_key", "chroma_key_vectorized"]
+__all__ = ["chroma_key"]
